src/main/java/com/ebm/domain/Team.java
package com.ebm.domain;

import java.time.LocalDateTime;
import java.util.List;

public class Team {

    private Integer id;
    private String name;
    private LocalDateTime dateJoin;


    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public LocalDateTime getDateJoin() {
        return dateJoin;
    }

    public void setDateJoin(LocalDateTime dateJoin) {
        this.dateJoin = dateJoin;
    }


}
